Organising Groups in folders to MVC like user groups
As Xcode shows the files in group and its really organised by all the
file are actually stored in the actual project folder so, viewing the
files in a normal file system other than Xcode looks very messy and
unorganised. I have created new folders with the same name as our
different groups like Model, View, Controller, JsonParser and set the
groups file location to the folders. So, now everything is organised in
MVC Patterns